pyproject.toml
[project]
name = "pseudo-entertainment"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=0.3.23",
    "langgraph>=0.5.0",
    "langchain-community>=0.2.17",
    "python-dotenv>=1.0.1",
    "langchain-openai>=0.3.12",
    "langchain_mcp_adapters>=0.1.9",
    "fastmcp<2.0.0",
    "mcp<2.0.0",
    "httpx>=0.24.0",
    "newsapi-python>=0.2.6",
    "langgraph-checkpoint-sqlite",
    "langchain-google-genai>=2.0.0",
]

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
    "langgraph-cli[inmem]>=0.2.3",
    "pre-commit>=4.2.0",
    "pytest>=8.3.5",
]

[tool.pyright]
include = ["agents", "tests"]
exclude = ["**/__pycache__/*", "**/node_modules/*", ".venv", "**/.venv"]

[tool.uv.workspace]
members = ["agents/text", "agents/music", "agents/image", "agents/management"]

[tool.setuptools.packages.find]
include = ["agents*"]

[tool.ruff]
exclude = [
    "(^|/)(.venv|venv|node_modules|__pycache__|migrations|static|templates|.git|.github|.langgraph_api|media)(/|$)",
    "^tests(/|$)",
    "^unit_tests(/|$)",
    ".env",
    ".env.example",
    ".gitignore",
    "LICENSE",
    "README.md",
    "langgraph.json",
    "uv.lock"
]
